backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pymongo==4.10.1
python-dotenv==1.0.1
scikit-learn==1.6.1
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
xgboost==2.1.3
joblib==1.4.2
shap==0.46.0
networkx==3.4.2
httpx==0.27.2
pydantic==2.10.4
slowapi==0.1.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiohttp==3.9.1
apscheduler==3.10.4
mistralai==1.2.5